// web/admin/src/types/api/v1/models/ProductResponse.ts
export interface ProductResponse {
    /**
     * 
     * @type {Category}
     * @memberof ProductResponse
     */
    category: Category;
    /**
     * 
     * @type {Coordinator}
     * @memberof ProductResponse
     */
    coordinator: Coordinator;
    /**
     * 
     * @type {Producer}
     * @memberof ProductResponse
     */
    producer: Producer;
    /**
     * 
     * @type {Product}
     * @memberof ProductResponse
     */
    product: Product;
    /**
     * 商品タグ一覧
     * @type {Array<ProductTag>}
     * @memberof ProductResponse
     */
    productTags: Array<ProductTag>;
    /**
     * 
     * @type {ProductType}
     * @memberof ProductResponse
     */
    productType: ProductType;
}